+++ CHANGES.rst
@@ -46,6 +46,7 @@ Improvements
 - Add CLI commands to block and unblock users (:issue:`3845`)
 - Show warning when trying to merge a blocked user (:issue:`3845`)
 - Allow importing event role members from a CSV file (:issue:`4301`)
+- Allow optional comment when accepting a pre-booking (:issue:`4086`)
 
 Bugfixes
 ^^^^^^^^